% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA2x.R
\name{NA2x}
\alias{NA2x}
\title{Replace numeric missing values with mean of the column}
\usage{
NA2x(data, verbose = FALSE, fun = mean, trim = 0)
}
\arguments{
\item{data}{a dataframe containing numeric variables (e.g double or integer).No default data selected, must be an argument.}

\item{verbose}{the verbose argument indicates the output was successful, by default verbose = FALSE as this function was designed for small data sets. For larger data sets, verbose = TRUE may be useful to know the function is being applied to  large data frames as the function will take longer. If TRUE, returns Processing successful.}

\item{fun}{specifies the function to be applied to missing values. By default, fun refers to mean(.x). An alternative function can be specified (fun = f(x)).}

\item{trim}{a fraction of outlying observations to be ignored when calculating for the replacement of NA. By default the trim is set to 0.}
}
\value{
With or without a specified function (default = mean), NA2x replaces each missing value with the mean of that column.
}
\description{
Replace numeric missing values with mean of the column
}
\examples{
```
# Create a data frame with both numeric and character variables.
df <-tibble(
 first = c("a", NA, "b", "c","d"),
 second = c(NA, 2, NA, 3, 4),
 third = c(10, NA, NA, 5, 6)
)


# Input data frame with missing values.
NA2x(df)

# The output is the same data frame with modified values.
# A tibble: 5 x 3
first     second third
<chr>      <dbl> <dbl>
1 a          3    10
2 NA         2     7
3 b          3     7
4 c          3     5
5 d          4     6
```
}
